+++ docs/CHANGELOG.md
@@ -1,7 +1,9 @@
 # ElasticsearchRecord - CHANGELOG
 
-## [1.0.0] - 2022-10-17
+## [1.0.0] - 2022-10-18
 * [add] patch for ```ActiveRecord::Relation::Merger``` - to support AR-relations
+* [add] ```ElasticsearchRecord::Relation#pit_results``` _offset_ & _yield_ support
+* [add] individual instrumentation names for ```ElasticsearchRecord::Relation``` result methods (like: #aggregations)
 * [rem] cleanup Debugging & logging
 * [fix] quoting for any values
 * [fix] calculation ```count``` method to support already known syntax (with column, distinct, limited, ...)
@@ -14,7 +16,9 @@
   * directly fail if a grouping _(visit_Arel_Nodes_Grouping)_ was provided
   * forced _failed!_ state was not correctly claimed
   * buggy _assign_-method for nested arrays
-* [fix] ```Arel::Collectors::ElasticsearchQuery``` claim for _argument_
+* [fix] ```Arel::Collectors::ElasticsearchQuery```
+  * claim for _argument_
+  * delete key on _nil_ assign
 * [ref] simplify ```Arel::Collectors::ElasticsearchQuery``` (remove stack & scoping)
 * [ref] simplify ```Arel::Visitors::Elasticsearch``` to support binds (statement cache) and simple where predicates
 * [ref] rename ```ElasticsearchRecord::Query#arguments``` -> _#query_arguments_